Acharya_Shreenidhi_A1/codefiles/requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.4
platformdirs==4.2.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
urllib3==2.2.0
waitress==2.1.2
Werkzeug==3.0.1
zeep==4.2.1
